backend/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6
httpx==0.26.0
requests==2.31.0
pillow==10.2.0
pydantic==2.5.3
pydantic-settings==2.1.0
python-dotenv==1.0.0

# Optional: Only needed if using Gemini
google-generativeai==0.3.2

# Vector Database (Pinecone)
pinecone-client==3.0.0

# FREE Local Embeddings (no API needed!)
sentence-transformers==2.3.1

# Supabase (PostgreSQL database)
supabase==2.3.0
